mobilesearch/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.4.0
djangorestframework==3.11.0
djongo==1.3.3
dnspython==1.16.0
gunicorn==20.0.4
httpie==2.2.0
idna==2.10
joblib==0.16.0
numpy==1.19.0
pickle-mixin==1.0.2
Pygments==2.6.1
pymongo==3.10.1
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.1
sklearn==0.0
sqlparse==0.2.4
threadpoolctl==2.1.0
urllib3==1.25.9
whitenoise==5.1.0
